+++ src/ui/activities/auth_activity/mod.rs
@@ -27,6 +27,7 @@
  */
 // Sub modules
 mod bookmarks;
+mod misc;
 mod update;
 mod view;